requirements.txt
# dùng bản python 3.7 nếu cài thư viện faiss bằng pip

absl-py==0.15.0
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
cached-property==1.5.2
cachetools==5.1.0
charset-normalizer==2.0.12
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
dask==2021.10.0
decorator==5.1.0
executing==0.8.3
# faiss==1.7.2
faiss-cpu==1.7.2
Flask==2.0.2
flatbuffers==1.12
fonttools==4.33.3
fsspec==2021.10.1
gast==0.3.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==3.3
imageio==2.10.1
importlib-metadata==4.11.4
ipython
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.2
libclang==14.0.1
locket==0.2.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
networkx==2.6.3
numpy
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.5.4.58
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.2
pandas
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.2
ply==3.11
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
six==1.15.0
sklearn==0.0
smop==0.30
stack-data==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tifffile==2021.10.12
tomli==2.0.1
toolz==0.11.1
tornado==3.2.1
tqdm==4.64.0
traitlets==5.2.1.post0
typed-ast==1.5.3
typing-extensions==3.7.4.3
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.8.0
